Updated for LyX 2.1, created variables, i8n
Former template would not produce PDF due to odd errors, seemingly due to apacite. Switching to another style (authordate1) is a workaround.

This version also cleans up the LaTeX preamble to only include newly-defined variables (the equivalent of [Course], etc.) so that they are defined once in the preamble and automatically appear in the headers/first page, etc. Removing the other things from the preamble was also beneficial (necessary) for me to make my document a French document (with Babel). LyX 2.1 supports renaming the entries such as Abstract/TOC/etc. to the translations of those in different languages. Note that some of the template text is still not completely language neutral, "e.g., course notes"  but they require small changes.
For hyperref I didn't bother changing the colors to match the ones set in the initial template. This is a configuration in LyX; see http://tex.stackexchange.com/a/82534/17868